Add event listener to change views
