+++ src/js/contact.js
@@ -1,12 +1,7 @@
 import { mainContent } from '../js/home';
-import insta from '../assets/instagram-icon.png';
-import face from '../assets/facebook-icon.png';
-import twitter from '../assets/twitter-icon.png';
-import mail from '../assets/email-icon.png';
 
 const elContact = document.createElement('div');
-const footerContact = document.querySelector('.contact-footer');
-export { elContact, footerContact, contact, contactFooter };
+export { elContact, contact };
 
 const contact = () => {
   elContact.classList.add('contact', 'hidden');
@@ -57,34 +52,4 @@ const contact = () => {
   formSubmitBtn.setAttribute("type", "submit");
   formSubmitBtn.setAttribute("value", "Submit");
   boxForm.appendChild(formSubmitBtn);
-}
-
-const contactFooter = () => {
-  const elFooter = document.createElement('div');
-  elFooter.classList.add('contact-footer','hidden');
-  mainContent.appendChild(elFooter);
-
-  const footerBox = document.createElement('div');
-  footerBox.classList.add('contact-footer-box', 'd-flex');
-  elFooter.appendChild(footerBox);
-
-  const boxItem1 = new Image();
-  boxItem1.src = insta;
-  boxItem1.classList.add('contact-footer-box-item');
-  footerBox.appendChild(boxItem1);
-
-  const boxItem2 = new Image();
-  boxItem2.src = face;
-  boxItem2.classList.add('contact-footer-box-item2');
-  footerBox.appendChild(boxItem2);
-
-  const boxItem3 = new Image();
-  boxItem3.src = twitter;
-  boxItem3.classList.add('contact-footer-box-item2');
-  footerBox.appendChild(boxItem3);
-
-  const boxItem4 = new Image();
-  boxItem4.src = mail;
-  boxItem4.classList.add('contact-footer-box-item');
-  footerBox.appendChild(boxItem4);
 }
