+++ src/js/menu.js
@@ -7,14 +7,9 @@ import jumbo from '../assets/crispyjumbo.png';
 import burguer1 from '../assets/crispyburguer1.png';
 import burguer2 from '../assets/crispyburguer2.png';
 import burguer3 from '../assets/crispyburguer3.png';
-import insta from '../assets/instagram-icon.png';
-import face from '../assets/facebook-icon.png';
-import twitter from '../assets/twitter-icon.png';
-import mail from '../assets/email-icon.png';
 
 const elMenu = document.createElement('div');
-const footerMenu = document.querySelector('.menu-footer');
-export { elMenu, footerMenu, menu, menuFooter };
+export { elMenu, menu };
 
 const menu = () => {
   elMenu.classList.add('menu', 'hidden');
@@ -151,34 +146,4 @@ const menu = () => {
   item8Price.classList.add('menu-box-item8-price');
   item8Price.innerHTML = '$17';
   boxItem8.appendChild(item8Price);
-}
-
-const menuFooter = () => {
-  const elFooter = document.createElement('div');
-  elFooter.classList.add('menu-footer','hidden');
-  mainContent.appendChild(elFooter);
-
-  const footerBox = document.createElement('div');
-  footerBox.classList.add('menu-footer-box', 'd-flex');
-  elFooter.appendChild(footerBox);
-
-  const boxItem1 = new Image();
-  boxItem1.src = insta;
-  boxItem1.classList.add('menu-footer-box-item');
-  footerBox.appendChild(boxItem1);
-
-  const boxItem2 = new Image();
-  boxItem2.src = face;
-  boxItem2.classList.add('menu-footer-box-item2');
-  footerBox.appendChild(boxItem2);
-
-  const boxItem3 = new Image();
-  boxItem3.src = twitter;
-  boxItem3.classList.add('menu-footer-box-item2');
-  footerBox.appendChild(boxItem3);
-
-  const boxItem4 = new Image();
-  boxItem4.src = mail;
-  boxItem4.classList.add('menu-footer-box-item');
-  footerBox.appendChild(boxItem4);
 }
